+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.5.6-dev+20.9.2014
 
+- Added the `textToScreenCoordinates` method to `VirtualRenderer`.
+
 ## 0.5.5+20.9.2014
 
 - Updated to ace-builds version 20.09.2014.